Add unification feature (as Prolog meaning) and context feature. An example about unification:
  %token a "|'
  %token b \w+
  #foo:
      ::a:: <b> ::a::

It will match "bar" and 'bar', but also "bar' and 'bar".
That's why we introduce unification. It works like this:

  %token a "|'
  %token b \w+
  #foo:
      ::a[0]:: <b> ::a[0]::

Index between square brackets must be greater or equal to zero but the number
does not matter, it is just for your convenience. It means that all the ::a[i]::
must be equals. It is the principle of unification :-). In this way, "bar" and
'bar' will be valid but "bar' and 'bar" will be invalid.