New arguments support.
Realist domains could define the $_arguments attributes, which could be null if
the realist domain has no argument, or “…” (see \Hoa\Core constants) to
an undefined arity, or finally an array to name arguments, e.g.:
    $_arguments = array('foo', 'bar'); : new …('a', 'b'); -> foo = a, bar = b.
    $_arguments = …;                   : new …('a', 'b', 'c'); -> list.
    $_arguments = null; (or nothing)   : new …();
To get, set, unset and check arguments, we use an array access on the realist
domain, e.g.: $this['foo'], $this['foo'] = 'baz', unset($this['foo']) and
isset($this['foo']). By default, undefined arguments are set to null.