Introduce choices when resolving components.
Hoa\Core\Protocol considers that every component defines a reach value
from the parent reach value: so relative value, no more absolute value.
In this way, a reach value could contain a “;” (semi-colon) character
that represents a choice. For example: 'dirA;dirB;dirC'. When resolving,
we compute all the possible paths and cache them, and after, we search
the first one that exists. When a reachId exists, combination are
interrupted.