Fix a segfault on PHP 5.5-dev and libxml 2.7.*.
Since SimpleXML has been updated to libxml 2.9.0 (pleas, see the commit
c5a611b25708d709ad5c539ed764e2a268fd04a7 on php-src.git) and if we work
with a lower version of libxml, a segfault appears because the
__toString() method is recursive (but it worked before and I don't know
why). I don't really understand the problem but it works with PHP 5.3.6
to 5.5-dev with libxml 2.7.*, so… :-).