Add support of \r in reach values.
When resolving a hoa:// path, we iterate over components, from the top
to the bottom. Current reach value is concatenated to reach value of
the parent. It is more efficient in most of the cases, but sometimes, we
might want to remove parent reach values, i.e. remove previous
concatenations.

To achieve that, we can use the \r (ASCII 13) character. It has exactly
the same semantics as defined in the POSIX standard. All characters at
the left of \r are removed. Thus:

    Hoa\Core::getInstance()->setProtocol(
        'hoa://Data/Temporary',
        "\r/tmp/"
    );
    var_dump(resolve('hoa://Data/Temporary')); // string(5) "/tmp/"

It is no more concatenated to hoa://Data reach value.

It also works with choices: "foo\rbar;baz\rqux", and the \r can be
everywhere you want (even if it has no meaning to not be the first
character).