Extend Composer support to hoa:// protocol.
First, we define the WITH_COMPOSER constant which is a flag set to true
if the core is loaded through Composer, else false. Consequently, we update
Hoa\Core\Consistency to use this constant.

Second, we create the Hoa\Core\Protocol\Library class to handle
components in hoa:// that refer to a library. This component will format
top-namespace and library name (respectively vendor and package with
Composer vocabulary) to lowercase if Composer is used.

Finally, we use this class when a component is called “Library”, such as:
hoa://Library or hoa://Data/Library.