Resolution can be stopped before finding choice.
When resolving a path, we collect choices. Then, we loop over these
solutions to find the first that exists. But in some cases, we would
like to only resolve the path, selecting the first choice without
checking if it represents an existing resource or not (e.g. if we would
like to compute a target to copy files). That is why we have introduced
the second argument: $exists to check whether or not we loop over
solutions.