Welcome to Hoa\Http\Runtime \o/.
This class allows to get informations about the current HTTP request. It
does not parse a request (it is the job of Hoa\Http\Request), it just
retrieves informations about the request that is responsible to the
current run.

For example:

    from('Hoa')
    -> import('Http.Runtime');

    var_dump(Hoa\Http\Runtime::getMethod());
    var_dump(Hoa\Http\Runtime::getData());
    var_dump(Hoa\Http\Runtime::getHeaders());

And:

    $ curl -X POST -d 'a=b&c=d' $uri
    string(4) "post"
    array(2) {
      ["a"]=>
      string(1) "b"
      ["c"]=>
      string(1) "d"
    }
    array(…) {
      …
      ["content-length"]=>
      string(1) "7"
      ["content-type"]=>
      string(33) "application/x-www-form-urlencoded"
      …
    }

Note: getData() with POST and Content-type: application/json decodes
JSON data.