Basic becomes a namespace & welcome to Controller.
With the Basic process, we instantiate the kit two times if a kit is
inherited by a controller. It is a wrong conception. We create
Basic\Controller to fill this role and it receives the kit in its
constructor. It is now easier to create its own controller helpers or
even dispatchers because it is namespaces-separated.