Fix a bug when importing with another load flag.
When we import a class with the load flag set to false, and then, we
import the same class again but with the load flag set to true, the
cache system ignores the new state of load flag. This is a bug. This
patch fixes that with some other API documentation typos.