New realistic domain API.
General
-------

Almost all realistic domains have been rewritten to fit the new API or to
use new features, such as constraints.

Auto-boxing
-----------

Hoa\Realdom has an auto-boxing system to transform on-the-fly PHP types
to realistic domains, thanks to const*() realistic domains.

Arguments
---------

It is possible to declare default values for arguments:

    protected $arguments = array(
        'foo' => 'bar',
        'baz' => 42
    );

These values will be auto-boxed.
It is possible to define types for arguments:

    protected $arguments = array(
        'String        foo' => 'bar',
        'Integer|Float baz' => 42
    );

Note it is possible to defines multiple types with the help of the pipe
(“|”) symbol.
It is possible to use the default value of an argument with the
“default” keyword in Praspel (not in PHP) and also with the “…” keyword
in Praspel and PHP.

These are three new exceptions related to argument:
    * Exception\InvalidArgument, if a given argument is invalid;
    * Exception\MissingArgument, if an argument is missing (not given
      and no default value);
    * Exception\InvalidArgument, if an argument is said invalid from the
      realistic domain, e.g. boundinteger(5, 2) will throw such an
      exception because 5 is not lower than 2.

Constraints
-----------

It is possible to declare constraints on realistic domains through
Praspel. These contraints are used to guide the sampling for example.