Introduce the @yield-select special attribute.
The @yield-select attribute allows to select attributes from the
original component and place it on the real component (in a yield@name).
For example, let use a <player /> component:

    <player id="foo" />

Its definition could be:

    <yield name="player">
      <div>
        …
      </div>
    </yield>

The HTML result (for example) would be:

    <div>
      …
    </div>

No @id="foo"? Embarrassing. That's why we have the @yield-select
attribute. See the new definition:

    <yield name="player">
      <div yield-select="?x:@*">
        …
      </div>
    </yield>

And the result is finally:

    <div id="foo">
      …
    </div>

Excellent :-).