Prevent a null node (please, see details).
Imagine the following scenario: a client and a server, merged together
(thanks to Hoa\Socket\Connection\Group) and running. The client send all
received data with the server. The server send all received data with
the client. If the client does not receive data immediately (before the
server), everything works. But, if the client receive a data before the
server has one connection up, i.e. one node up, we have a null node, so
an error. Fix it by not sending the data :-).