Hoa\Core\Event::unregister can be soft or hard.
Before, when unregistering an event, both source and attached callables
were deleted. Now, by default, only the source is deleted. In this way,
we profit of the “asynchroneous registration process”. We are able to
register again a new source with the same attached callables.

If we would like to delete both source and callables, we must set the
second argument to true. It is the (previous and) hard way.