Preload is not required by default.
Remove \ArrayAccess interface and remove the three previous autoloader
states. Now we have only one parameter that matters: whether preload is
required or not (thanks to the methods requirePreload() and
isPreloadRequired()). By default, the preload is NOT required.

Thus, we can use the famous from/import calls to preload classes (with
all the benefits: import pattern, such as * or ~, aliases for entry
classes, foreachImport(), etc.) or we can use the autoloader directly. In
all cases, we can use the `use` keyword obviously.