Remove reference when setting the value.
If the variable is an argument that is passed by-reference, and, if its
value is changed during the execution of the system, its reference will
be propagated until here. Consequently, in the post-state of the system,
the value of the variable in the pre-state will be equal to the value in
the post-state. This is an error. Thus, we copy the value instead of
using a reference (initially introduced for performance reason).