Fix condition when using abstract uri.
When we use a custom router [like here](http://hoa-project.net/En/Literature/Learn/Routerdispatcher.html), we can requests on uri who `resolve` can't find.

So when BHoa check if the target exists, the return is false (line 232) but just after (line 329) BHoa do a `pathinfo` on this target (who it's not a file) and check if the extension is `php`.

But like we can see on [php.net](http://www.php.net/manual/en/function.pathinfo.php): *If the path does not have an extension, no extension element will be returned*.

So Hoa throw a error and we finish with a fatal error uncaught exception.

I don't understand how the [example](http://hoa-project.net/En/Literature/Learn/Routerdispatcher.html) can work without that... or I've miss something.