Decode URI after having splitted subdomain & URI.
We can use `@` to specify a subdomain and a URI (for instance
`sub@/uri/`) in the same `$uri`.

Previously, we decoded URI before splitting the subdomain and the URI.
Consequently, if a URI contains `%40`, being `@`, everything at the left
of `%40` will be considered as a subdomain, and therefore, it is highly
likely to fail the routing.

Now we decode URI after having splitted the subdomain and the URI.