Force MIME type with the constructor.
For instance, if the stream is a string buffer, `Hoa\Mime` will throw an
exception because the stream name has no extension so the MIME type
cannot be identified. With this patch, we can force the MIME type. It
also avoids to instanciate `Hoa\Mime`, which saves some CPU and memory.