Avoid sending a second `EHLO` if no `STARTTLS`.
In the case the server does not support `STARTTLS`, we always re-sent a
new `EHLO`. While this operation is mandatory (see RFC3207, Section 4.2)
if we switched to TLS, it is totally useless in the case no `STARTTLS`
command has been sent. So move the new `EHLO` in the `STARTTLS` block.
It will speed mail sending in the case `STARTTLS` is not supported by
the server.