Update retro-compatibility with PHP7.
The new exception architecture introduced in PHP7 [1] has been totally
redesigned [2]. This patch updates the retro-compatibility classes
according to this new architecture. Consequently, the `BaseException`
class has been removed, along with `EngineException` and
`ParseException`. While these latters could be implemented (not as is),
we prefer to, so far, only implement the `Throwable` interface. Let see
if we can implement (still for the retro-compatibility) the `Error`,
`TypeError` and `ParseError` class.

[1]: https://wiki.php.net/rfc/engine_exceptions_for_php7
[2]: https://wiki.php.net/rfc/throwable-interface