Transport: Introduce vendor schemes.
Now we are able to register our own transports, like: `ws://foo/bar`.

The idea is very simple: A pair $(t, c)$ is registered on the
`Hoa\Socket\Transport` class where $t$ is a transport scheme (like `ws`)
and $c$ is a callable. This callable is a factory and is used to
build a valid `Hoa\Socket\Socket` instance based on an URI (like
`foo/bar`).