Introduce the `pause` and `resume` methods.
In [issue 29], for really short execution, the global mark was longer
than 100% of the time, which is not possible. This is due to the fact we
take time between each mark computation when drawing the statistics.

Consequently, the `pause` and `resume` methods are introduced to
respectively pause all runnings tasks and resume a specific set of
marks.

We use these methods in the `getStatistic` method to pause all the
running tasks and resume them before and after actually computing the
statistics.

[issue 29]: https://github.com/hoaproject/Bench/issues/29