Only compute the documentation once.
If we find twice the same library, it was computed twice. This caused an
issue. For instance, if we are inside a library that has a
documentation, it will be scanned twice: First time because we scan the
current library and second time because we scan `hoa://Library`.

This patch fixes this issue by using the hashmap unicity property of PHP
arrays and by reformatting the `$vendors` array for XYL.