Documentation: Fix schema loading error.
By default, if no remote or referrer is present for a
central.hoa-project.net resource, we will fallback to Github.
Unfortunately, Github adds a `X-Frame-Options: deny` header
([X-Frame-Options], [RFC7034]). So we force the remote to `hoa` to solve
this problem since `git.hoa-project.net` does not add this header. Also,
we can trust our own services ;-).

[X-Frame-Options]: https://developer.mozilla.org/en-US/docs/Web/HTTP/X-Frame-Options
[RFC7034]: http://tools.ietf.org/html/rfc7034