VFS: Force `type=file`. If absent, no resolution.
Previously, if no `type` query was specify, we considered `type=file` by
default. Unfortunately, this was therefore impossible to check whether a
file exists or not because it was always created.

Now, we use `?type=xxx` to force the creation of a file (`file` or
`directory`). If absent, we check whether it exists or not. This is more
consistent.