Test: Reverse a `while` to a `do`/`while`.
The `run` method does a `while(true)`. This is impossible to test. So we
reverse it for a `do { … } while(true)` and we mock the `true` constant
to `false` for only one pass in the loop. However, we use `SUCCEED`
instead of `true`, which is a synonym, but a bug in some PHP versions
does not allow to mock `true` correctly (this is a constant but with a
special treatement apparently).