Grammar: Support UTF-8 strings.
UTF-8 is an issue with JSON, because we must handle surrogate pairs
([1], [2]). This patch implements UTF-8 not only for validating a datum
but also for generating a datum. It means that we only generate valid
UTF-8 strings and we only validate/recognize valid UTF-8 strings.

UTF-16 strings will follow in another patch.

Unfortunately, we decided to implement the whole string as a token, not
as rules. This is unfortunate because grammar coverage algorithms in
`Hoa\Compiler` applies only on rules, not on tokens. This is a potential
optimisation.

[1]: http://tools.ietf.org/html/rfc7159#section-7
[2]: http://tools.ietf.org/html/rfc7159#section-8