Server: Fix throws exceptions while broadcasting messages.
While broadcasting a message to some nodes an error can occur because the node connection is closed but isn't yet removed from the *to send* list.
Here we will isolate each sends and then thrown a group of BrokenPipeException. This way if the first node is closed each other will still receive the message and then we'll emit an exception.