Generator: Support `ignore` in code block type.
Each code block must has the type `php`. However, the type is a list of
verbs, separated by a comma. So `php,ignore` is valid, and contains two
verbs: `php`, and `ignore`.

If the verb `ignore` is present, then the code block is compiled into a
test case but it is marked as `skip`.