Generator: Support `must_throw` in code block type.
Just type `ignore`, `must_throw` is a new verb that can be used in a
code block type. Thus `php,must_throw` represents an code block that is
a PHP program but it must throw an exception.

So far, exceptions must be of kind `Exception`. It restricts exceptions
to the runtime exception category.