Generator: Support `use` and `#`.
Code block can have comments. If the comment starts with `#` (shell
style), then the whole comment is removed for the HTML API browser, but
they are kept when compiling examples into test cases. So the following
example:

     # $a = 1;
     $b = 2;
     assert(3 === $a + $b);

will be compiled as the following test case:

     $a = 1;
     $b = 2;
     assert(3 === $a + $b);

and will be displayed as follows for the HTML version:

     $b = 2;
     assert(3 === $a + $b);

This is useful when we would like to add `use` statements in
comments. In our context, we cannot use `use` because test cases are
methods, and the syntax of PHP does not allow `use` statements inside
methods. So we must expanded `use` statements to remove them.

To address that, the new `unfoldCode` method lexes and expands `use`
statements. For instance:

    # use Foo\Bar;
    new Bar\Baz();

will be unfolded as:

    # use Foo\Bar;
    new \Foo\Bar\Baz();

To achieve this, first, the comments are removed, and, second, `<?php`
is prepended:

    <?php
    use Foo\Bar;
    new Bar\Baz();

Then, third, we use the `token_get_all` native lexer to lex the code
block, and rewrite it. Finally, when rewriting, the `use` statements are
put in comments (of kinds `#`), even if they were not in comments
before. The `<?php` opening tag is removed too. The result is:

    # use Foo\Bar;
    new \Foo\Bar\Baz();